move payload with modalities to top.py
