+++ forte/data/data_pack.py
@@ -46,7 +46,6 @@ from forte.data.ontology.core import Entry, Grid
 from forte.data.ontology.core import EntryType
 from forte.data.ontology.top import (
     Annotation,
-    Grids,
     Link,
     Group,
     SinglePackEntries,
@@ -172,7 +171,6 @@ class DataPack(BasePack[Entry, Link, Group]):
         self.image_annotations: List[ImageAnnotation] = []
         self.grid: Optional[Grid] = None
 
-        self.grids: List[Grids] = []
         self.text_payloads: List[Payload] = []
         self.audio_payloads: List[Payload] = []
         self.image_payloads: List[Payload] = []
