+++ ft/onto/base_ontology.py
@@ -54,9 +54,6 @@ __all__ = [
     "MRCQuestion",
     "Recording",
     "AudioUtterance",
-    "AudioPayload",
-    "TextPayload",
-    "ImagePayload",
 ]
 
 
@@ -561,37 +558,3 @@ class AudioUtterance(AudioAnnotation):
         super().__init__(pack, begin, end)
         self.speaker: Optional[str] = None
 
-
-@dataclass
-class AudioPayload(Payload):
-    """
-    A payload that caches audio data
-    Attributes:
-        sample_rate (Optional[int]):
-    """
-
-    sample_rate: Optional[int]
-
-    def __init__(self, pack: DataPack, payload_idx: int = 0, uri: Optional[str] = None):
-        super().__init__(pack, payload_idx, uri)
-        self.sample_rate: Optional[int] = None
-
-
-@dataclass
-class TextPayload(Payload):
-    """
-    A payload that caches text data
-    """
-
-    def __init__(self, pack: DataPack, payload_idx: int = 0, uri: Optional[str] = None):
-        super().__init__(pack, payload_idx, uri)
-
-
-@dataclass
-class ImagePayload(Payload):
-    """
-    A payload that caches image data
-    """
-
-    def __init__(self, pack: DataPack, payload_idx: int = 0, uri: Optional[str] = None):
-        super().__init__(pack, payload_idx, uri)
